cli_blockquote {cli}	R Documentation

CLI block quote

Description

A section that is quoted from another source. It is typically indented.

Usage

cli_blockquote(
  quote,
  citation = NULL,
  id = NULL,
  class = NULL,
  .envir = parent.frame()
)

Arguments

quote
Text of the quotation.

citation
Source of the quotation, typically a link or the name of a person.

id
Element id, a string. If NULL, then a new id is generated and returned.

class
Class name, sting. Can be used in themes.

.envir
Environment to evaluate the glue expressions in. It is also used to auto-close the container if .auto_close is TRUE.

Examples

cli_blockquote(cli:::lorem_ipsum(), citation = "Nobody, ever")

[Package cli version 2.5.0 Index]