cli_code {cli}	R Documentation

A block of code

Description

A helper function that creates a div with class code and then calls cli_verbatim() to output code lines. The builtin theme formats these containers specially. In particular, it adds syntax highlighting to valid R code.

Usage

cli_code(
  lines = NULL,
  ...,
  language = "R",
  .auto_close = TRUE,
  .envir = environment()
)

Arguments

lines
Character vector, each line will be a line of code, and newline characters also create new lines. Note that no glue substitution is performed on the code.

...
More character vectors, they are appended to lines.

language
Programming language. This is also added as a class, in addition to code.

.auto_close
Passed to cli_div() when creating the container of the code. By default the code container is closed after emitting lines and ... via cli_verbatim(). You can keep that container open with .auto_close and/or .envir, and then calling cli_verbatim() to add (more) code. Note that the code will be formatted and syntax highlighted separately for each cli_verbatim() call.

.envir
Passed to cli_div() when creating the container of the code.

Details

myfun <- function() {
  message("Just an example function")
  graphics::pairs(iris, col = 1:4)
}
cli_code(format(myfun))

#> function ()                                                                     
#> {                                                                               
#>     message("Just an example function")                                         
#>     graphics::pairs(iris, col = 1:4)                                            
#> }                                                                               

Value

The id of the container that contains the code.

[Package cli version 3.6.1 Index]