Definition list

A definition list is a container, see containers.

cli_dl(
  items = NULL,
  id = NULL,
  class = NULL,
  .close = TRUE,
  .auto_close = TRUE,
  .envir = parent.frame()
)

Arguments

items
Named character vector, or NULL. If not NULL, they are used as list items.

id
Id of the list container. Can be used for closing it with cli_end() or in themes. If NULL, then an id is generated and retuned invisibly.

class
Class of the list container. Can be used in themes.

.close
Whether to close the list container if the items were specified. If FALSE then new items can be added to the list.

.auto_close
Whether to close the container, when the calling function finishes (or .envir is removed, if specified).

.envir
Environment to evaluate the glue expressions in. It is also used to auto-close the container if .auto_close is TRUE.

Value

The id of the new container element, invisibly.

Examples

## Specifying the items at the beginning
cli_dl(c(foo = "one", bar = "two", baz = "three"))

#> foo: one

#> bar: two

#> baz: three

## Adding items one by one
cli_dl()
cli_li(c(foo = "one"))

#> foo: one

cli_li(c(bar = "two"))

#> bar: two

cli_li(c(baz = "three"))

#> baz: three

cli_end()