cli_list_themes {cli}	R Documentation

List the currently active themes

Description

If there is no active app, then it calls start_app().

Usage

cli_list_themes()

Value

A list of data frames with the active themes. Each data frame row is a style that applies to selected CLI tree nodes. Each data frame has columns:

selector: The original CSS-like selector string. See themes.

parsed: The parsed selector, as used by cli for matching to nodes.

style: The original style.

cnt: The id of the container the style is currently applied to, or NA if the style is not used.

See Also

themes

[Package cli version 2.5.0 Index]