Ordered CLI list

An ordered list is a container, see containers.

cli_ol(
  items = NULL,
  id = NULL,
  class = NULL,
  .close = TRUE,
  .auto_close = TRUE,
  .envir = parent.frame()
)

Arguments

items
If not NULL, then a character vector. Each element of the vector will be one list item, and the list container will be closed by default (see the .close argument).

id
Id of the list container. Can be used for closing it with cli_end() or in themes. If NULL, then an id is generated and retuned invisibly.

class
Class of the list container. Can be used in themes.

.close
Whether to close the list container if the items were specified. If FALSE then new items can be added to the list.

.auto_close
Whether to close the container, when the calling function finishes (or .envir is removed, if specified).

.envir
Environment to evaluate the glue expressions in. It is also used to auto-close the container if .auto_close is TRUE.

Value

The id of the new container element, invisibly.

Examples

## Specifying the items at the beginning
cli_ol(c("one", "two", "three"))

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   1. one

#>   2. two

#>   3. three

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

## Adding items one by one
cli_ol()

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

cli_li("one")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   1. one

cli_li("two")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   2. two

cli_li("three")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   3. three

cli_end()

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

## Nested lists
cli_div(theme = list(ol = list("margin-left" = 2)))

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

cli_ul()

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

cli_li("one")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>     ─ one

cli_ol(c("foo", "bar", "foobar"))

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>     1. foo

#>     2. bar

#>     3. foobar

cli_li("two")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>     ─ two

cli_end()

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

cli_end()

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149