cli_output_connection {cli}	R Documentation

The connection option that cli would use

Description

Note that this only refers to the current R process. If the output is produced in another process, then it is not relevant.

Usage

cli_output_connection()

Details

In interactive sessions the standard output is chosen, othrwise the standard error is used. This is to avoid painting output messages red in the R GUIs.

Value

Connection object.

[Package cli version 2.5.0 Index]