format_inline {cli}	R Documentation

Format and returns a line of text

Description

You can use this function to format a line of cli text, without emitting it to the screen. It uses cli_text() internally.

Usage

format_inline(
  ...,
  .envir = parent.frame(),
  collapse = TRUE,
  keep_whitespace = TRUE
)

Arguments

...
Passed to cli_text().

.envir
Environment to evaluate the expressions in.

collapse
Whether to collapse the result if it has multiple lines, e.g. because of \f characters.

keep_whitespace
Whether to keep all whitepace (spaces, newlines and form feeds) as is in the input.

Details

format_inline() performs no width-wrapping.

Value

Character scalar, the formatted string.

See Also

This function supports inline markup.

Other functions supporting inline markup: cli_abort(), cli_alert(), cli_blockquote(), cli_bullets_raw(), cli_bullets(), cli_dl(), cli_h1(), cli_li(), cli_ol(), cli_process_start(), cli_progress_along(), cli_progress_bar(), cli_progress_message(), cli_progress_output(), cli_progress_step(), cli_rule, cli_status_update(), cli_status(), cli_text(), cli_ul(), format_error()

Examples

format_inline("A message for {.emph later}, thanks {.fn format_inline}.")

[Package cli version 3.6.1 Index]