Character vector to put a spinner on the screen

cli contains many different spinners, you choose one according to your taste.

get_spinner(which = NULL)

Arguments

which
The name of the chosen spinner. The default depends on whether the platform supports Unicode.

Value

A list with entries: name, interval: the suggested update interval in milliseconds and frames: the character vector of the spinner's frames.

See also

Examples

get_spinner()

#> $name
#> [1] "dots"
#> 
#> $interval
#> [1] 80
#> 
#> $frames
#>  [1] "⠋" "⠙" "⠹" "⠸" "⠼" "⠴" "⠦" "⠧" "⠇" "⠏"
#> 

get_spinner("shark")

#> $name
#> [1] "shark"
#> 
#> $interval
#> [1] 120
#> 
#> $frames
#>  [1] "▐|\\____________▌" "▐_|\\___________▌" "▐__|\\__________▌"
#>  [4] "▐___|\\_________▌" "▐____|\\________▌" "▐_____|\\_______▌"
#>  [7] "▐______|\\______▌" "▐_______|\\_____▌" "▐________|\\____▌"
#> [10] "▐_________|\\___▌" "▐__________|\\__▌" "▐___________|\\_▌"
#> [13] "▐____________|\\▌" "▐____________/|▌"  "▐___________/|_▌" 
#> [16] "▐__________/|__▌"  "▐_________/|___▌"  "▐________/|____▌" 
#> [19] "▐_______/|_____▌"  "▐______/|______▌"  "▐_____/|_______▌" 
#> [22] "▐____/|________▌"  "▐___/|_________▌"  "▐__/|__________▌" 
#> [25] "▐_/|___________▌"  "▐/|____________▌" 
#>