Progress bar variables

Progress bar variables

Details

These variables can be used in cli progress bar format strings.

pb_bar creates a visual progress bar. If the number of total units is unknown, then it will return an empty string.

pb_current is the number of current progress units.

pb_current_bytes is the number of current progress utils formatted as bytes. The output has a constant width of six characters.

pb_elapsed is the elapsed time since the start of the progress bar. The time is measured since the progress bar is created with cli_progress_bar() or similar.

pb_elapsed_clock is the elapsed time, in hh::mm::ss format.

pb_elapsed_raw is the number of seconds since the start of the progress bar.

pb_eta is the estimated time until the end of the progress bar, in human readable form.

pb_eta_raw is the estimated time until the end of the progress bar, in seconds.

pb_eta_str is the estimated time until the end of the progress bar. It includes the "ETA:" prefix. It is only shown if the time can be estimated, otherwise it is the empty string.

pb_extra can be used to access extra data, see the extra argument of cli_progress_bar() and cli_progress_update().

pb_id is the id of the progress bar. The id has the format cli-<pid>-<counter> where <pid> is the process id, and <counter> is an integer counter that is incremented every time cli needs a new unique id.

pb_name is the name of the progress bar. This is supplied by the developer, and it is by default the empty string. A space character is added to non-empty names.

pb_percent is the percentage of the progress bar, always formatted in three characters plus the percentage sign. If the total number of units is unknown, then it is " NA%".

pb_pid is the integer process id of the progress bar.

pb_rate is the progress rate, in number of units per second, formatted in a string.

pb_rate_raw is the raw progress rate, in number of units per second.

pb_rate_bytes is the progress rate, formatted as bytes per second, in human readable form.

pb_spin is a spinner. The default spinner is selected via a get_spinner() call.

pb_status is the status string of the progress bar. By default this is an empty string, but it is possible to set it in cli_progress_bar() and `cli_progress_update()].

pb_timestamp is a time stamp in ISO 8601 format.

pb_total is the total number of progress units, or NA if the number of units is unknown.

pb_total_bytes is the total number of progress units, formatted as bytes, in a human readable format.

Examples

# pb_bar and pb_percent
cli_progress_demo(
  format = "Progress bar: {cli::pb_bar} {cli::pb_percent}",
  total = 100
)
#> Progress bar: ■                                  1%
#> Progress bar: ■■                                 2%
#> Progress bar: ■■                                 3%
#> Progress bar: ■■                                 4%
#> Progress bar: ■■■                                5%
#> Progress bar: ■■■                                6%
#> Progress bar: ■■■                                7%
#> Progress bar: ■■■                                8%
#> Progress bar: ■■■■                               9%
#> Progress bar: ■■■■                              10%
#> Progress bar: ■■■■                              11%
#> Progress bar: ■■■■■                             12%
#> Progress bar: ■■■■■                             13%
#> Progress bar: ■■■■■                             14%
#> Progress bar: ■■■■■                             15%
#> Progress bar: ■■■■■■                            16%
#> Progress bar: ■■■■■■                            17%
#> Progress bar: ■■■■■■                            18%
#> Progress bar: ■■■■■■■                           19%
#> Progress bar: ■■■■■■■                           20%
#> Progress bar: ■■■■■■■                           21%
#> Progress bar: ■■■■■■■■                          22%
#> Progress bar: ■■■■■■■■                          23%
#> Progress bar: ■■■■■■■■                          24%
#> Progress bar: ■■■■■■■■■                         25%
#> Progress bar: ■■■■■■■■■                         26%
#> Progress bar: ■■■■■■■■■                         27%
#> Progress bar: ■■■■■■■■■                         28%
#> Progress bar: ■■■■■■■■■■                        29%
#> Progress bar: ■■■■■■■■■■                        30%
#> Progress bar: ■■■■■■■■■■                        31%
#> Progress bar: ■■■■■■■■■■■                       32%
#> Progress bar: ■■■■■■■■■■■                       33%
#> Progress bar: ■■■■■■■■■■■                       34%
#> Progress bar: ■■■■■■■■■■■                       35%
#> Progress bar: ■■■■■■■■■■■■                      36%
#> Progress bar: ■■■■■■■■■■■■                      37%
#> Progress bar: ■■■■■■■■■■■■                      38%
#> Progress bar: ■■■■■■■■■■■■■                     39%
#> Progress bar: ■■■■■■■■■■■■■                     40%
#> Progress bar: ■■■■■■■■■■■■■                     41%
#> Progress bar: ■■■■■■■■■■■■■■                    42%
#> Progress bar: ■■■■■■■■■■■■■■                    43%
#> Progress bar: ■■■■■■■■■■■■■■                    44%
#> Progress bar: ■■■■■■■■■■■■■■■                   45%
#> Progress bar: ■■■■■■■■■■■■■■■                   46%
#> Progress bar: ■■■■■■■■■■■■■■■                   47%
#> Progress bar: ■■■■■■■■■■■■■■■                   48%
#> Progress bar: ■■■■■■■■■■■■■■■■                  49%
#> Progress bar: ■■■■■■■■■■■■■■■■                  50%
#> Progress bar: ■■■■■■■■■■■■■■■■                  51%
#> Progress bar: ■■■■■■■■■■■■■■■■■                 52%
#> Progress bar: ■■■■■■■■■■■■■■■■■                 53%
#> Progress bar: ■■■■■■■■■■■■■■■■■                 54%
#> Progress bar: ■■■■■■■■■■■■■■■■■                 55%
#> Progress bar: ■■■■■■■■■■■■■■■■■■                56%
#> Progress bar: ■■■■■■■■■■■■■■■■■■                57%
#> Progress bar: ■■■■■■■■■■■■■■■■■■                58%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■               59%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■               60%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■               61%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■              62%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■              63%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■              64%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■             65%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■             66%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■             67%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■             68%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■            69%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■            70%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■            71%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■           72%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■           73%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■           74%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■           75%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■          76%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■          77%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■          78%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■         79%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■         80%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■         81%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■        82%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■        83%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■        84%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■       85%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■       86%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■       88%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      89%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      90%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      91%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     92%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     93%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     94%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     95%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    96%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    97%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    98%
#> Progress bar: ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   99%
#> 

# pb_current and pb_total
cli_progress_demo(
  format = "[{cli::pb_current}/{cli::pb_total}]",
  total = 248
)
#> [1/248]
#> [2/248]
#> [3/248]
#> [4/248]
#> [5/248]
#> [6/248]
#> [7/248]
#> [8/248]
#> [9/248]
#> [10/248]
#> [11/248]
#> [12/248]
#> [13/248]
#> [14/248]
#> [15/248]
#> [16/248]
#> [17/248]
#> [18/248]
#> [19/248]
#> [20/248]
#> [21/248]
#> [22/248]
#> [23/248]
#> [24/248]
#> [25/248]
#> [26/248]
#> [27/248]
#> [28/248]
#> [29/248]
#> [30/248]
#> [31/248]
#> [32/248]
#> [33/248]
#> [34/248]
#> [35/248]
#> [36/248]
#> [37/248]
#> [38/248]
#> [39/248]
#> [40/248]
#> [41/248]
#> [42/248]
#> [43/248]
#> [44/248]
#> [45/248]
#> [46/248]
#> [47/248]
#> [48/248]
#> [49/248]
#> [50/248]
#> [51/248]
#> [52/248]
#> [53/248]
#> [54/248]
#> [55/248]
#> [56/248]
#> [57/248]
#> [58/248]
#> [59/248]
#> [60/248]
#> [61/248]
#> [62/248]
#> [63/248]
#> [64/248]
#> [65/248]
#> [66/248]
#> [67/248]
#> [68/248]
#> [69/248]
#> [70/248]
#> [71/248]
#> [72/248]
#> [73/248]
#> [74/248]
#> [75/248]
#> [76/248]
#> [77/248]
#> [78/248]
#> [79/248]
#> [80/248]
#> [81/248]
#> [82/248]
#> [83/248]
#> [84/248]
#> [85/248]
#> [86/248]
#> [87/248]
#> [88/248]
#> [89/248]
#> [90/248]
#> [91/248]
#> [92/248]
#> [93/248]
#> [94/248]
#> [95/248]
#> [96/248]
#> [97/248]
#> [98/248]
#> [99/248]
#> [100/248]
#> [101/248]
#> [102/248]
#> [103/248]
#> [104/248]
#> [105/248]
#> [106/248]
#> [107/248]
#> [108/248]
#> [109/248]
#> [110/248]
#> [111/248]
#> [112/248]
#> [113/248]
#> [114/248]
#> [115/248]
#> [116/248]
#> [117/248]
#> [118/248]
#> [119/248]
#> [120/248]
#> [121/248]
#> [122/248]
#> [123/248]
#> [124/248]
#> [125/248]
#> [126/248]
#> [127/248]
#> [128/248]
#> [129/248]
#> [130/248]
#> [131/248]
#> [132/248]
#> [133/248]
#> [134/248]
#> [135/248]
#> [136/248]
#> [137/248]
#> [138/248]
#> [139/248]
#> [140/248]
#> [141/248]
#> [142/248]
#> [143/248]
#> [144/248]
#> [145/248]
#> [146/248]
#> [147/248]
#> [148/248]
#> [149/248]
#> [150/248]
#> [151/248]
#> [152/248]
#> [153/248]
#> [154/248]
#> [155/248]
#> [156/248]
#> [157/248]
#> [158/248]
#> [159/248]
#> [160/248]
#> [161/248]
#> [162/248]
#> [163/248]
#> [164/248]
#> [165/248]
#> [166/248]
#> [167/248]
#> [168/248]
#> [169/248]
#> [170/248]
#> [171/248]
#> [172/248]
#> [173/248]
#> [174/248]
#> [175/248]
#> [176/248]
#> [177/248]
#> [178/248]
#> [179/248]
#> [180/248]
#> [181/248]
#> [182/248]
#> [183/248]
#> [184/248]
#> [185/248]
#> [186/248]
#> [187/248]
#> [188/248]
#> [189/248]
#> [190/248]
#> [191/248]
#> [192/248]
#> [193/248]
#> [194/248]
#> [195/248]
#> [196/248]
#> [197/248]
#> [198/248]
#> [199/248]
#> [200/248]
#> [201/248]
#> [202/248]
#> [203/248]
#> [204/248]
#> [205/248]
#> [206/248]
#> [207/248]
#> [208/248]
#> [209/248]
#> [210/248]
#> [211/248]
#> [212/248]
#> [213/248]
#> [214/248]
#> [215/248]
#> [216/248]
#> [217/248]
#> [218/248]
#> [219/248]
#> [220/248]
#> [221/248]
#> [222/248]
#> [223/248]
#> [224/248]
#> [225/248]
#> [226/248]
#> [227/248]
#> [228/248]
#> [229/248]
#> [230/248]
#> [231/248]
#> [232/248]
#> [233/248]
#> [234/248]
#> [235/248]
#> [236/248]
#> [237/248]
#> [238/248]
#> [239/248]
#> [240/248]
#> [241/248]
#> [242/248]
#> [243/248]
#> [244/248]
#> [245/248]
#> [246/248]
#> [247/248]
#> 

# pb_current_bytes, pb_total_bytes
cli_progress_demo(
  format = "[{cli::pb_current_bytes}/{cli::pb_total_bytes}]",
  total = 102800,
  at = seq(0, 102800, by = 1024)
)
#> [0.0 kB/103 kB]
#> [1.0 kB/103 kB]
#> [2.0 kB/103 kB]
#> [3.1 kB/103 kB]
#> [4.1 kB/103 kB]
#> [5.1 kB/103 kB]
#> [6.1 kB/103 kB]
#> [7.2 kB/103 kB]
#> [8.2 kB/103 kB]
#> [9.2 kB/103 kB]
#> [ 10 kB/103 kB]
#> [ 11 kB/103 kB]
#> [ 12 kB/103 kB]
#> [ 13 kB/103 kB]
#> [ 14 kB/103 kB]
#> [ 15 kB/103 kB]
#> [ 16 kB/103 kB]
#> [ 17 kB/103 kB]
#> [ 18 kB/103 kB]
#> [ 19 kB/103 kB]
#> [ 20 kB/103 kB]
#> [ 22 kB/103 kB]
#> [ 23 kB/103 kB]
#> [ 24 kB/103 kB]
#> [ 25 kB/103 kB]
#> [ 26 kB/103 kB]
#> [ 27 kB/103 kB]
#> [ 28 kB/103 kB]
#> [ 29 kB/103 kB]
#> [ 30 kB/103 kB]
#> [ 31 kB/103 kB]
#> [ 32 kB/103 kB]
#> [ 33 kB/103 kB]
#> [ 34 kB/103 kB]
#> [ 35 kB/103 kB]
#> [ 36 kB/103 kB]
#> [ 37 kB/103 kB]
#> [ 38 kB/103 kB]
#> [ 39 kB/103 kB]
#> [ 40 kB/103 kB]
#> [ 41 kB/103 kB]
#> [ 42 kB/103 kB]
#> [ 43 kB/103 kB]
#> [ 44 kB/103 kB]
#> [ 45 kB/103 kB]
#> [ 46 kB/103 kB]
#> [ 47 kB/103 kB]
#> [ 48 kB/103 kB]
#> [ 49 kB/103 kB]
#> [ 50 kB/103 kB]
#> [ 51 kB/103 kB]
#> [ 52 kB/103 kB]
#> [ 53 kB/103 kB]
#> [ 54 kB/103 kB]
#> [ 55 kB/103 kB]
#> [ 56 kB/103 kB]
#> [ 57 kB/103 kB]
#> [ 58 kB/103 kB]
#> [ 59 kB/103 kB]
#> [ 60 kB/103 kB]
#> [ 61 kB/103 kB]
#> [ 62 kB/103 kB]
#> [ 63 kB/103 kB]
#> [ 65 kB/103 kB]
#> [ 66 kB/103 kB]
#> [ 67 kB/103 kB]
#> [ 68 kB/103 kB]
#> [ 69 kB/103 kB]
#> [ 70 kB/103 kB]
#> [ 71 kB/103 kB]
#> [ 72 kB/103 kB]
#> [ 73 kB/103 kB]
#> [ 74 kB/103 kB]
#> [ 75 kB/103 kB]
#> [ 76 kB/103 kB]
#> [ 77 kB/103 kB]
#> [ 78 kB/103 kB]
#> [ 79 kB/103 kB]
#> [ 80 kB/103 kB]
#> [ 81 kB/103 kB]
#> [ 82 kB/103 kB]
#> [ 83 kB/103 kB]
#> [ 84 kB/103 kB]
#> [ 85 kB/103 kB]
#> [ 86 kB/103 kB]
#> [ 87 kB/103 kB]
#> [ 88 kB/103 kB]
#> [ 89 kB/103 kB]
#> [ 90 kB/103 kB]
#> [ 91 kB/103 kB]
#> [ 92 kB/103 kB]
#> [ 93 kB/103 kB]
#> [ 94 kB/103 kB]
#> [ 95 kB/103 kB]
#> [ 96 kB/103 kB]
#> [ 97 kB/103 kB]
#> [ 98 kB/103 kB]
#> [ 99 kB/103 kB]
#> [100 kB/103 kB]
#> [101 kB/103 kB]
#> [102 kB/103 kB]