spark_bar {cli}	R Documentation

Draw a sparkline bar graph with unicode block characters

Description

Rendered using block elements. In most common fixed width fonts these are rendered wider than regular characters which means they are not suitable if you need precise alignment.

Usage

spark_bar(x)

Arguments

x
A numeric vector between 0 and 1

Details

You might want to avoid sparklines on non-UTF-8 systems, because they do not look good. You can use is_utf8_output() to test for support for them.

See Also

spark_line()

Examples

x <- seq(0, 1, length = 6)
spark_bar(x)
spark_bar(sample(x))

spark_bar(seq(0, 1, length = 8))

# NAs are left out
spark_bar(c(0, NA, 0.5, NA, 1))

[Package cli version 2.5.0 Index]