symbol {cli}	R Documentation

Various handy symbols to use in a command line UI

Description

Various handy symbols to use in a command line UI

Usage

symbol

list_symbols()

Format

A named list, see names(symbol) for all sign names.

Details

On Windows they have a fallback to less fancy symbols.

list_symbols() prints a table with all symbols to the screen.

Examples

cat(symbol$tick, " SUCCESS\n", symbol$cross, " FAILURE\n", sep = "")

## All symbols
cat(paste(format(names(symbol), width = 20),
  unlist(symbol)), sep = "\n")

[Package cli version 2.5.0 Index]