Wrap an ANSI styled string to a certain width

This function is similar to base::strwrap(), but works on ANSI styled strings, and leaves the styling intact.

Usage

ansi_strwrap(
  x,
  width = console_width(),
  indent = 0,
  exdent = 0,
  simplify = TRUE
)

Arguments

x

ANSI string.

width

Width to wrap to.

indent

Indentation of the first line of each paragraph.

exdent

Indentation of the subsequent lines of each paragraph.

simplify

Whether to return all wrapped strings in a single character vector, or wrap each element of x independently and return a list.

Value

If simplify is FALSE, then a list of character vectors, each an ANSI string. Otherwise a single ANSI string vector.

See also

Examples

text <- cli:::lorem_ipsum()
# Highlight some words, that start with 's'
rexp <- gregexpr("\\b([sS][a-zA-Z]+)\\b", text)
regmatches(text, rexp) <- lapply(regmatches(text, rexp), col_red)
cat(text)
#> Laboris duis nostrud amet excepteur est officia. Incididunt consequat dolor dolor nisi in pariatur consectetur officia. Lorem ad non laborum labore in voluptate nulla quis do ad aliqua reprehenderit quis. Mollit labore esse ad elit. Et dolore ex sint duis enim voluptate excepteur adipisicing sint ea. Voluptate quis cupidatat ea sit officia quis sit elit fugiat. Est magna pariatur nisi reprehenderit fugiat. Sunt sint ullamco veniam ea sunt id ex dolor laborum sint est pariatur aliqua occaecat. Qui nostrud Lorem ex enim laborum nostrud amet eu aliqua ex irure laborum et dolore. Ex sunt officia quis eu eiusmod irure qui non Lorem cupidatat enim irure.

wrp <- ansi_strwrap(text, width = 40)
cat(wrp, sep = "\n")
#> Laboris duis nostrud amet excepteur est
#> officia. Incididunt consequat dolor
#> dolor nisi in pariatur consectetur
#> officia. Lorem ad non laborum labore in
#> voluptate nulla quis do ad aliqua
#> reprehenderit quis. Mollit labore esse
#> ad elit. Et dolore ex sint duis enim
#> voluptate excepteur adipisicing sint
#> ea. Voluptate quis cupidatat ea sit
#> officia quis sit elit fugiat. Est magna
#> pariatur nisi reprehenderit fugiat.
#> Sunt sint ullamco veniam ea sunt id ex
#> dolor laborum sint est pariatur aliqua
#> occaecat. Qui nostrud Lorem ex enim
#> laborum nostrud amet eu aliqua ex irure
#> laborum et dolore. Ex sunt officia quis
#> eu eiusmod irure qui non Lorem
#> cupidatat enim irure.