Wrap an ANSI styled string to a certain width

This function is similar to base::strwrap(), but works on ANSI styled strings, and leaves the styling intact.

Usage

ansi_strwrap(
  x,
  width = console_width(),
  indent = 0,
  exdent = 0,
  simplify = TRUE
)

Arguments

x

ANSI string.

width

Width to wrap to.

indent

Indentation of the first line of each paragraph.

exdent

Indentation of the subsequent lines of each paragraph.

simplify

Whether to return all wrapped strings in a single character vector, or wrap each element of x independently and return a list.

Value

If simplify is FALSE, then a list of character vectors, each an ANSI string. Otherwise a single ANSI string vector.

See also

Examples

text <- cli:::lorem_ipsum()
# Highlight some words, that start with 's'
rexp <- gregexpr("\\b([sS][a-zA-Z]+)\\b", text)
regmatches(text, rexp) <- lapply(regmatches(text, rexp), col_red)
cat(text)
#> Non quis eu laboris fugiat esse laboris consectetur ut. Amet velit non anim do culpa id veniam. Elit et incididunt aliquip nisi ea eu aute in minim duis. Deserunt pariatur proident et velit ut consequat quis eiusmod occaecat ea veniam fugiat occaecat in. Dolore incididunt non id do officia in nostrud nostrud consectetur ea sint sint ex excepteur. Quis consectetur id mollit occaecat Lorem voluptate nostrud nisi irure commodo labore. Id aliqua reprehenderit dolore minim in ea aliqua consectetur. Eiusmod sint anim adipisicing tempor eu eu sit magna eiusmod veniam eu.

wrp <- ansi_strwrap(text, width = 40)
cat(wrp, sep = "\n")
#> Non quis eu laboris fugiat esse laboris
#> consectetur ut. Amet velit non anim do
#> culpa id veniam. Elit et incididunt
#> aliquip nisi ea eu aute in minim duis.
#> Deserunt pariatur proident et velit ut
#> consequat quis eiusmod occaecat ea
#> veniam fugiat occaecat in. Dolore
#> incididunt non id do officia in nostrud
#> nostrud consectetur ea sint sint ex
#> excepteur. Quis consectetur id mollit
#> occaecat Lorem voluptate nostrud nisi
#> irure commodo labore. Id aliqua
#> reprehenderit dolore minim in ea aliqua
#> consectetur. Eiusmod sint anim
#> adipisicing tempor eu eu sit magna
#> eiusmod veniam eu.