builtin_theme {cli}	R Documentation

The built-in CLI theme

Description

This theme is always active, and it is at the bottom of the theme stack. See themes.

Usage

builtin_theme(dark = getOption("cli_theme_dark", "auto"))

Arguments

dark
Whether to use a dark theme. The cli_theme_dark option can be used to request a dark theme explicitly. If this is not set, or set to "auto", then cli tries to detect a dark theme, this works in recent RStudio versions and in iTerm on macOS.

Value

A named list, a CLI theme.

See Also

themes, simple_theme().

[Package cli version 2.5.0 Index]