cli: Helpers for Developing Command Line Interfaces

A suite of tools to build attractive command line interfaces ('CLIs'), from semantic elements: headings, lists, alerts, paragraphs, etc. Supports custom themes via a 'CSS'-like language. It also contains a number of lower level 'CLI' elements: rules, boxes, trees, and 'Unicode' symbols with 'ASCII' alternatives. It support ANSI colors and text styles as well.

See also

Author

Maintainer: Gábor Csárdi gabor@posit.co

Other contributors:

Hadley Wickham [contributor]

Kirill Müller [contributor]

Salim Brüggemann salim-b@pm.me (ORCID) [contributor]

Posit Software, PBC [copyright holder, funder]