cli-package {cli}	R Documentation

cli: Helpers for Developing Command Line Interfaces

Description

A suite of tools to build attractive command line interfaces ('CLIs'), from semantic elements: headings, lists, alerts, paragraphs, etc. Supports custom themes via a 'CSS'-like language. It also contains a number of lower level 'CLI' elements: rules, boxes, trees, and 'Unicode' symbols with 'ASCII' alternatives. It support ANSI colors and text styles as well.

Author(s)

Maintainer: Gábor Csárdi csardi.gabor@gmail.com

Other contributors:

Hadley Wickham [contributor]

Kirill Müller [contributor]

RStudio [copyright holder, funder]

See Also

Useful links:

https://cli.r-lib.org

https://github.com/r-lib/cli#readme

Report bugs at https://github.com/r-lib/cli/issues

[Package cli version 3.6.1 Index]