Debug cli internals

Return the current state of a clie app. It includes the currencly open tags, their ids, classes and their computed styles.

cli_debug_doc(app = default_app() %||% start_app())

Arguments

app
The cli app to debug. Defaults to the current app. if there is no app, then it creates one by calling start_app().

Value

Data frame with columns: tag, id, class (space separated), theme (id of the theme the element added), styles (computed styles for the element).

Details

The returned data frame has a print method, and if you want to create a plain data frame from it, index it with an empty bracket: cli_debug_doc()[].

To see all currently active themes, use app$themes, e.g. for the default app: default_app()$themes.

See also

Examples

if (FALSE) {
cli_debug_doc()

olid <- cli_ol()
cli_li()
cli_debug_doc()
cli_debug_doc()[]

cli_end(olid)
cli_debug_doc()
}