cli_format {cli}	R Documentation

Format a value for printing

Description

This function can be used directly, or via the {.val ...} inline style. {.val {expr}} calls cli_format() automatically on the value of expr, before styling and collapsing it.

Usage

cli_format(x, style = NULL, ...)

## Default S3 method:
cli_format(x, style = NULL, ...)

## S3 method for class 'character'
cli_format(x, style = NULL, ...)

## S3 method for class 'numeric'
cli_format(x, style = NULL, ...)

Arguments

x
The object to format.

style
List of formatting options, see the individual methods for the style options they support.

...
Additional arguments for methods.

Details

It is possible to define new S3 methods for cli_format and then these will be used automatically for {.val ...} expressions.

See Also

cli_vec()

Examples

things <- c(rep("this", 3), "that")
cli_format(things)
cli_text("{.val {things}}")

nums <- 1:5 / 7
cli_format(nums, style = list(digits = 2))
cli_text("{.val {nums}}")
divid <- cli_div(theme = list(.val = list(digits = 3)))
cli_text("{.val {nums}}")
cli_end(divid)

[Package cli version 2.5.0 Index]