cli_li {cli}	R Documentation

CLI list item(s)

Description

A list item is a container, see containers.

Usage

cli_li(
  items = NULL,
  id = NULL,
  class = NULL,
  .auto_close = TRUE,
  .envir = parent.frame()
)

Arguments

items
Character vector of items, or NULL.

id
Id of the new container. Can be used for closing it with cli_end() or in themes. If NULL, then an id is generated and returned invisibly.

class
Class of the item container. Can be used in themes.

.auto_close
Whether to close the container, when the calling function finishes (or .envir is removed, if specified).

.envir
Environment to evaluate the glue expressions in. It is also used to auto-close the container if .auto_close is TRUE.

Value

The id of the new container element, invisibly.

Examples

## Adding items one by one
cli_ul()
cli_li("one")
cli_li("two")
cli_li("three")
cli_end()

## Complex item, added gradually.
cli_ul()
cli_li()
cli_verbatim("Beginning of the {.emph first} item")
cli_text("Still the first item")
cli_end()
cli_li("Second item")
cli_end()

[Package cli version 2.5.0 Index]