List of built-in cli progress styles

The following options are used to select a style:

cli_progress_bar_style

cli_progress_bar_style_ascii

cli_progress_bar_style_unicode

Usage

cli_progress_styles()

Value

A named list with sublists containing elements complete, incomplete and potentially current.

Details

On Unicode terminals (if is_utf8_output() is TRUE), the cli_progress_bar_style_unicode and cli_progress_bar_style options are used.

On ASCII terminals (if is_utf8_output() is FALSE), the cli_pgoress_bar_style_ascii and cli_progress_bar_style options are are used.

for (style in names(cli_progress_styles())) {
  options(cli.progress_bar_style = style)
  label <- ansi_align(paste0("Style '", style, "'"), 20)
  print(cli_progress_demo(label, live = FALSE, at = 66, total = 100))
}
options(cli.progress_var_style = NULL)

#> Style 'classic'      #####################             66% | ETA:  3s
#> Style 'squares'      ■■■■■■■■■■■■■■■■■■■■■             66% | ETA:  3s
#> Style 'dot'          ────────────────────●──────────   66% | ETA:  3s
#> Style 'fillsquares'  ■■■■■■■■■■■■■■■■■■■■■□□□□□□□□□□   66% | ETA:  3s
#> Style 'bar'          ███████████████████████████████   66% | ETA:  3s

See also