Update the status bar

The status bar is the last line of the terminal. cli apps can use this to show status information, progress bars, etc. The status bar is kept intact by all semantic cli output.

`cli_status() creates a status.

cli_status_update() updates the contents of an existing status.

cli_status_clear() terminates a status.

cli_status(
  msg,
  msg_done = paste(msg, "... done"),
  msg_failed = paste(msg, "... failed"),
  .keep = FALSE,
  .auto_close = TRUE,
  .envir = parent.frame(),
  .auto_result = c("autoclear", "clear", "done", "failed", "auto")
)

cli_status_clear(
  id = NULL,
  result = c("clear", "done", "failed"),
  msg_done = NULL,
  msg_failed = NULL,
  .envir = parent.frame()
)

cli_status_update(
  id = NULL,
  msg = NULL,
  msg_done = NULL,
  msg_failed = NULL,
  .envir = parent.frame()
)

Arguments

msg
The text to show, a character vector. It will be collapsed into a single string, and the first line is kept and cut to console_width(). The message is often associated with the start of a calculation. You can use cli inline markup in this message. If the message contains glue interpolation, then it will be evaluated every time the message is updated. cli_status_update() may update this message.

msg_done
The message to use when the status is terminated, if the calculation finishes successfully. If .auto_close is TRUE and .auto_result is "done", then this is printed automatically when the calling function (or .envir) finishes. You can use cli inline markup in this message. If the message contains glue interpolation, then it be evaluated only when the status terminates. cli_status_update() may update this message.

msg_failed
The message to use when the status terminates, when the calculation finishes unsuccessfully. If .auto_close is TRUE and .auto_result is "failed", then this is printed automatically when the calling function (or .envir) finishes. You can use cli inline markup in this message. If the message contains glue interpolation, then it be evaluated only when the status terminates. cli_status_update() may update this message.

.keep
This argument is now ignored, and only kept for compatibility. Use the .auto_result argument to specify how to auto-terminate the status. Or use the result argument of cli_status_clear() if you clear the status manually.

.auto_close
Whether to terminate the status when the calling function finishes (or ‘.envir’ is removed from thae stack, if specified).

.envir
Environment to evaluate the glue expressions in. It is also used to auto-clear the status if .auto_close is TRUE.

.auto_result
What to do when auto-terminating the status. Possible values:

clear: The status is removed from the status bar.

done: The status is removed from the status bar, and the msg_done message is printed.

failed: The status is removed from the status bar, and the msg_failed message is printed.

auto: Automatically choose between done, for regular termination, and failed, for R errors and interruptions.

autoclear (default): Automatically choose between clear, for regular termination, and failed, for R errors and interruptions.

id
Id of the status to update or terminate. If it is NULL (the default), then the current status is manipulated. If cli_status_update() cannot find a status with id (or id is NULL, and there is no current status), then a warning is issued.

result
The type of termination, may be "clear", "done" or "failed". See more at the .auto_result parameter and below.

Value

cli_status returns the id of the new status, invisibly.

cli_status_clear returns NULL.

cli_status_update() returns the id of the status.

Details

The life of a cli status

You create a cli status with cli_status() or cli_process_start(). cli will show the new status in the status bar.

If neccessary, you update a status with cli_status_update() or cli_process_update().

You terminate a status with cli_status_clear(), cli_process_done() or cli_process_failed(). When a status is terminated, it is removed from the status bar. Before that, it may still print a final summary message, which you can specify using the msg_done and msg_failed arguments.

cli can also auto-terminate a status, see 'Automatic termination' below.

A status has three modes of termination:

clear: The status is removed from the status bar.

done: The status is removed from the status bar, but its msg_done message is printed as a regular cli message. This is typically a short summary of a task that has just finished, e.g. Downloaded 14 files (1.5MB)."

failed: The status is removed from the status bar, and its msg_failed message is printed as a regular cli message. (For auto-termination caused by an R error, this message typically shows up under R's error message.)

Status ids and the current status

When you create a status cli assigns an id to it. This id can be used to update or terminate the status.

If you don't specify the id in cli_status_update(), cli_process_update(), etc. then cli will manipulate the current status. The current status is the one that was last created in the caller (or in .envir if specified). This allows simpler code:

done <- 0
cli_process_start("Downloading data files: {done}/{total}")
for (done in seq_len(total)) {
  ...
  cli_process_update()
}

...

done <- 0
cli_process_start("Processing data files: {done}/{total}")
for (done in seq_len(total)) {
  ...
  cli_process_update()
}

Here the cli_process_update() calls refer to previoues cli_status() calls.

Also note that we didn't have to terminate the first status explicitly, thanks to cli's automatic termination, see the next section.

A similar example with the cli_status* API:

done <- 0
cli_status("Downloading data files: {done}/{total}")
for (done in seq_len(total)) {
  ...
  cli_status_update()
}

...

done <- 0
cli_status("Processing data files: {done}/{total}")
for (done in seq_len(total)) {
  ...
  cli_status_update()
}

Automatic termination

Usually you will want the status to terminate automatically if an error happens and/or the function that created it quits.

In most cases you also want to terminate the status if another status is created from the same called, via cli_status() or cli_process_start().

If the .auto_close argument of cli_status() and cli_process_start() is TRUE (the default), then cli will automatically terminate the status is these cases. This helps to reduce the clutter in the code caused by cli messages.

If you don't want to terminate a status automatically, you need to specify .auto_close = FALSE. In this case, we suggest that you add an on.exit() or withr::defer() handler that terminates the status, to avoid keeping it on the screen after an error or interruption happens.

Multiple active statuses

It is possible to have multiple active cli statuses at the same time. This may happen when you creates statues from nested function calls, or when you specify .auto_close = FALSE.

Nested statuses "just work" in general: if a nested level status terminates, cli will automatically show its immediate nesting status in the status bar.

Currently, when you update a status, cli will show it in the status bar, even if it there are other statuses nested "inside" it. This behavior might change in the future, as it can cause flickering when multiple statuses are updated frequently.

Differences between cli_status_* and cli_process_*

For historical reasons cli has two sets of functions to manage the status bar: the cli_status* functions and the cli_process_* functions.

Both sets are equivalent, but have different defaults:

cli_process_* prints termination messages by default.

cli_process_* by default styles the status messages, the same way as

cli_alert_info(), while the status is active,

cli_alert_success() for successful termination,

cli_alert_danger() for unsuccessful termination.

If you want to style the cli_status*() status messages, you need to add classes manually, e.g.

cli_status(msg = "{.alert Starting computation}")

cli_process_* omits the ... done and ... failed suffixes from the default termination messages, as the styling conveys this information.

See also