format_inline {cli}	R Documentation

Format and returns a line of text

Description

You can use this function to format a line of cli text, without emitting it to the screen. It uses cli_text() internally.

Usage

format_inline(..., .envir = parent.frame())

Arguments

...
Passed to cli_text().

.envir
Environment to evaluate the expressions in.

Details

format_inline() performs no width-wrapping.

Value

Character scalar, the formatted string.

Examples

format_inline("A message for {.emph later}, thanks {.fn format_inline}.")

[Package cli version 3.1.1 Index]