get_spinner {cli}	R Documentation

Character vector to put a spinner on the screen

Description

cli contains many different spinners, you choose one according to your taste.

Usage

get_spinner(which = NULL)

Arguments

which
The name of the chosen spinner. The default depends on whether the platform supports Unicode.

Value

A list with entries: name, interval: the suggested update interval in milliseconds and frames: the character vector of the spinner's frames.

See Also

Other spinners: demo_spinners(), list_spinners(), make_spinner()

Examples

get_spinner()
get_spinner("shark")

[Package cli version 2.5.0 Index]