CLI inline markup

CLI inline markup

Command substitution

All text emitted by cli supports glue interpolation. Expressions enclosed by braces will be evaluated as R code. See glue::glue() for details.

In addition to regular glue interpolation, cli can also add classes to parts of the text, and these classes can be used in themes. For example

cli_text("This is {.emph important}.")

adds a class to the "important" word, class "emph". Note that in this case the string within the braces is usually not a valid R expression. If you want to mix classes with interpolation, add another pair of braces:

adjective <- "great"
cli_text("This is {.emph {adjective}}.")

An inline class will always create a span element internally. So in themes, you can use the span.emph CSS selector to change how inline text is emphasized:

cli_div(theme = list(span.emph = list(color = "red")))
adjective <- "nice and red"
cli_text("This is {.emph {adjective}}.")

Classes

The default theme defines the following inline classes:

emph for emphasized text.

strong for strong importance.

code for a piece of code.

pkg for a package name.

fun for a function name.

arg for a function argument.

key for a keyboard key.

file for a file name.

path for a path (essentially the same as file).

email for an email address.

url for a URL.

var for a variable name.

envvar for the name of an environment variable.

See examples below.

You can simply add new classes by defining them in the theme, and then using them, see the example below.

Collapsing inline vectors

When cli performs inline text formatting, it automatically collapses glue substitutions, after formatting. This is handy to create lists of files, packages, etc. See examples below.

Escaping { and }

It might happen that you want to pass a string to cli_* functions, and you do not_ want command substitution in that string, because it might contain } and { characters. The simplest solution for this is referring to the string from a template:

msg <- "Error in if (ncol(dat$y)) {: argument is of length zero"
cli_alert_warning("{msg}")

If you want to explicitly escape { and } characters, just double them:

cli_alert_warning("A warning with {{ braces }}")

See also examples below.

Examples

## Some inline markup examples
cli_ul()

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149
#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

cli_li("{.emph Emphasized} text")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ Emphasized text

cli_li("{.strong Strong} importance")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ Strong importance

cli_li("A piece of code: {.code sum(a) / length(a)}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ A piece of code: `sum(a) / length(a)`

cli_li("A package name: {.pkg cli}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ A package name: cli

cli_li("A function name: {.fn cli_text}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ A function name: cli_text()

cli_li("A keyboard key: press {.kbd ENTER}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ A keyboard key: press <ENTER>

cli_li("A file name: {.file /usr/bin/env}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ A file name: /usr/bin/env

cli_li("An email address: {.email bugs.bunny@acme.com}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ An email address: bugs.bunny@acme.com

cli_li("A URL: {.url https://acme.com}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ A URL: <https://acme.com>

cli_li("An environment variable: {.envvar R_LIBS}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ─ An environment variable: R_LIBS

cli_end()

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

## Adding a new class
cli_div(theme = list(
  span.myclass = list(color = "lightgrey"),
  "span.myclass" = list(before = "["),
  "span.myclass" = list(after = "]")))

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

cli_text("This is {.myclass in brackets}.")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   This is [in brackets].

cli_end()

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

## Collapsing
pkgs <- c("pkg1", "pkg2", "pkg3")
cli_text("Packages: {pkgs}.")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   Packages: pkg1, pkg2 and pkg3.

cli_text("Packages: {.pkg {pkgs}}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   Packages: pkg1, pkg2 and pkg3

## Escaping
msg <- "Error in if (ncol(dat$y)) {: argument is of length zero"
cli_alert_warning("{msg}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ! Error in if (ncol(dat$y)) {: argument is of length zero

cli_alert_warning("A warning with {{ braces }}")

#> Error in mycall(sym_write, fd, data) : 
#>   Cannot write to fd (system error 9, Bad file descriptor) @client.c:149

#>   ! A warning with { braces }